+++ classes/util.php
@@ -58,10 +58,39 @@ class util {
     /** Force saving the document has failed. */
     const STATUS_ERRORFORCESAVE = 7;
 
+    /** Path locales to create file from ONLYOFFICE template. */
+    const PATH_LOCALE = [
+        "az" => "az-Latn-AZ",
+        "bg" => "bg-BG",
+        "cs" => "cs-CZ",
+        "de" => "de-DE",
+        "el" => "el-GR",
+        "en-GB" => "en-GB",
+        "en" => "en-US",
+        "es" => "es-ES",
+        "fr" => "fr-FR",
+        "gl" => "gl-ES",
+        "it" => "it-IT",
+        "ja" => "ja-JP",
+        "ko" => "ko-KR",
+        "lv" => "lv-LV",
+        "nl" => "nl-NL",
+        "pl" => "pl-PL",
+        "pt-BR" => "pt-BR",
+        "pt" => "pt-PT",
+        "ru" => "ru-RU",
+        "sk" => "sk-SK",
+        "sv" => "sv-SE",
+        "tr" => "tr-TR",
+        "uk" => "uk-UA",
+        "vi" => "vi-VN",
+        "zh" => "zh-CN"];
+
     /**
      * Get plugin key.
      *
      * @return string plugin key from the plugin configuration.
+     * @throws \dml_exception
      */
     public static function get_appkey() {
         $key = get_config('onlyofficeeditor', 'appkey');
@@ -155,6 +184,54 @@ class util {
         } else {
             return false;
         }
+    }
+
+    /**
+     * Create new empty file for ONLYOFFICE activity.
+     *
+     * @param string $fileformat new file format.
+     * @param object $user user.
+     * @param int $contextid context id.
+     * @param string $dirroot moodle dir root.
+     * @param int $fileid new file id.
+     * @param string $name name of the new file.
+     * @throws \file_exception
+     * @throws \stored_file_creation_exception
+     */
+    public static function create_from_onlyoffice_template($fileformat, $user, $contextid, $dirroot, $fileid, $name) {
+        switch ($fileformat) {
+            case 'Document': {
+                $fileformat = 'docx';
+                break;
+            }
+            case 'Spreadsheet': {
+                $fileformat = 'xlsx';
+                break;
+            }
+            case 'Presentation': {
+                $fileformat = 'pptx';
+                break;
+            }
+            case 'Form template': {
+                $fileformat = 'docxf';
+                break;
+            }
+        }
+        $pathlocale = self::PATH_LOCALE[$user->lang];
+        $pathname = $dirroot . '/mod/onlyofficeeditor/newdocs/' . $pathlocale . '/new.' . $fileformat;
+
+        $fileinfo = array(
+            'author' => fullname($user),
+            'contextid' => $contextid,
+            'component' => 'mod_onlyofficeeditor',
+            'filearea' => 'content',
+            'userid' => $user->id,
+            'itemid' => $fileid,
+            'filepath' => '/',
+            'filename' => $name . '.' . $fileformat);
+
+        $fs = get_file_storage();
+        $file = $fs->create_file_from_pathname($fileinfo, $pathname);
     }
 
 }